First refactor running
